Merge 8 years of zend parser changes into XHP parser.
Rather than trying to frankenstein new PHP features into a parser that was orked from Zend in 2011, apply all of the changes that have happened in PHP since then. In the process, I found a lot of syntax that ext_xhp didn't handle (and a few parser concepts that only mattered in PHP that Zend removed with the AST implementation). Overall this should make it easier to keep up with new syntax in the future, since the core grammar (minus XHP-specific syntax) should be identical.